noq_proto::crypto::rustls

Struct QuicServerConfig

pub struct QuicServerConfig { /* private fields */ }

A QUIC-compatible TLS server configuration

noq implicitly constructs a QuicServerConfig with reasonable defaults within ServerConfig::with_single_cert(). Alternatively, QuicServerConfig’s TryFrom implementation or with_initial method can be used to wrap around a custom [rustls::ServerConfig], in which case care should be taken around certain points:

If max_early_data_size is not set to u32::MAX, the server will not be able to accept incoming 0-RTT data. QUIC prohibits max_early_data_size values other than 0 or u32::MAX.

The rustls::ServerConfig must have TLS 1.3 support enabled for conversion to succeed.

Implementations

impl QuicServerConfig

pub fn with_initial( inner: Arc<ServerConfig>, initial: Suite, ) -> Result<Self, NoInitialCipherSuite>

Initialize a QUIC-compatible TLS client configuration with a separate initial cipher suite

This is useful if you want to avoid the initial cipher suite for traffic encryption.

pub fn set_alpn_protocols(&mut self, alpn_protocols: Vec<Vec<u8>>)

Updates the set of ALPN protocols configured in the server config.