noq_proto

Struct Connection

pub struct Connection { /* private fields */ }

Protocol state and logic for a single QUIC connection

Objects of this type receive ConnectionEvents and emit EndpointEvents and application Events to make progress. To handle timeouts, a Connection returns timer updates and expects timeouts through various methods. A number of simple getter methods are exposed to allow callers to inspect some of the connection state.

Connection has roughly 4 types of methods:

A. Simple getters, taking &self

B. Handlers for incoming events from the network or system, named handle_*.

C. State machine mutators, for incoming commands from the application. For convenience we refer to this as “performing I/O” below, however as per the design of this library none of the functions actually perform system-level I/O. For example, read and write, but also things like reset.

D. Polling functions for outgoing events or actions for the caller to take, named poll_*.

The simplest way to use this API correctly is to call (B) and (C) whenever appropriate, then after each of those calls, as soon as feasible call all polling methods (D) and deal with their outputs appropriately, e.g. by passing it to the application or by making a system-level I/O call. You should call the polling functions in this order:

poll_transmit

poll_timeout

poll_endpoint_events

poll

Currently the only actual dependency is from (2) to (1), however additional dependencies may be added in future, so the above order is recommended.

(A) may be called whenever desired.

Care should be made to ensure that the input events represent monotonically increasing time. Specifically, calling handle_timeout with events of the same Instant may be interleaved in any order with a call to handle_event at that same instant; however events or timeouts with different instants must not be interleaved.

Implementations

impl Connection

pub fn poll_timeout(&self) -> Option<Instant>

Returns the next time at which handle_timeout should be called

The value returned may change after:

the application performed some I/O on the connection

a call was made to handle_event

a call to poll_transmit returned Some

a call was made to handle_timeout

pub fn poll(&mut self) -> Option<Event>

Returns application-facing events

Connections should be polled for events after:

a call was made to handle_event

a call was made to handle_timeout

pub fn poll_endpoint_events(&mut self) -> Option<EndpointEvent>

Return endpoint-facing events

pub fn streams(&mut self) -> Streams<'_>

Provide control over streams

pub fn recv_stream(&mut self, id: StreamId) -> RecvStream<'_>

Provide control over streams

pub fn send_stream(&mut self, id: StreamId) -> SendStream<'_>

Provide control over streams

pub fn open_path_ensure( &mut self, network_path: FourTuple, initial_status: PathStatus, now: Instant, ) -> Result<(PathId, bool), PathError>

Opens a new path only if no path on the same network path currently exists.

Returns (path_id, true) if the path already existed, or (path_id, false) if was opened.

If network_path has no local IP set, then this will open a new path if no path exists for this remote address, independent of the existing path’s local IP. If a local IP is set, it will match against the full four-tuple of existing paths. Not setting the local IP avoids having to guess which local interface will be used to communicate with the remote, should it not be known yet. We assume that if we already have a path to the remote, the OS is likely to use the same interface to talk to said remote.

See also open_path.

pub fn open_path( &mut self, network_path: FourTuple, initial_status: PathStatus, now: Instant, ) -> Result<PathId, PathError>

Opens a new path.

Further errors might occur and they will be emitted in PathEvent::Abandoned events with this path id. Once the path is opened and can carry application data it will be reported using a PathEvent::Established event.

pub fn close_path( &mut self, now: Instant, path_id: PathId, error_code: VarInt, ) -> Result<(), ClosePathError>

Closes a path and sends a PATH_ABANDON frame with the passed error code.

Returns ClosePathError::LastOpenPath if this is the last open path. It does allow closing paths which have not yet been opened, as e.g. is the case when receiving a PATH_ABANDON from the peer for a path that was never opened locally.

pub fn paths(&self) -> Vec<PathId>

Returns all known paths.

There is no guarantee any of these paths are open or usable.

pub fn path_status(&self, path_id: PathId) -> Result<PathStatus, ClosedPath>

Gets the local PathStatus for a known PathId

pub fn network_path(&self, path_id: PathId) -> Result<FourTuple, ClosedPath>

Returns the path’s network path represented as a 4-tuple.

pub fn set_path_status( &mut self, path_id: PathId, status: PathStatus, ) -> Result<PathStatus, SetPathStatusError>

Sets the PathStatus for a known PathId

Returns the previous path status on success.

pub fn remote_path_status(&self, path_id: PathId) -> Option<PathStatus>

Returns the remote path status

pub fn set_path_max_idle_timeout( &mut self, now: Instant, path_id: PathId, timeout: Option<Duration>, ) -> Result<Option<Duration>, ClosedPath>

Sets the max_idle_timeout for a specific path.

If Some, the path idle timer is immediately re-armed. Setting None disables the timeout and stops the timer.

See TransportConfig::default_path_max_idle_timeout for details.

Returns the previous value of the setting.

pub fn set_path_keep_alive_interval( &mut self, path_id: PathId, interval: Option<Duration>, ) -> Result<Option<Duration>, ClosedPath>

Sets the keep_alive_interval for a specific path

See TransportConfig::default_path_keep_alive_interval for details.

Returns the previous value of the setting.

pub fn poll_transmit( &mut self, now: Instant, max_datagrams: NonZeroUsize, buf: &mut Vec<u8>, ) -> Option<Transmit>

Returns packets to transmit

Connections should be polled for transmit after:

the application performed some I/O on the connection

a call was made to handle_event

a call was made to handle_timeout

max_datagrams specifies how many datagrams can be returned inside a single Transmit using GSO. This must be at least 1.

pub fn handle_event(&mut self, event: ConnectionEvent)

Process ConnectionEvents generated by the associated Endpoint

Will execute protocol logic upon receipt of a connection event, in turn preparing signals (including application Events, EndpointEvents and outgoing datagrams) that should be extracted through the relevant methods.

pub fn handle_timeout(&mut self, now: Instant)

Process timer expirations

Executes protocol logic, potentially preparing signals (including application Events, EndpointEvents and outgoing datagrams) that should be extracted through the relevant methods.

It is most efficient to call this immediately after the system clock reaches the latest Instant that was output by poll_timeout; however spurious extra calls will simply no-op and therefore are safe.

pub fn close(&mut self, now: Instant, error_code: VarInt, reason: Bytes)

Close a connection immediately

This does not ensure delivery of outstanding data. It is the application’s responsibility to call this only when all important communications have been completed, e.g. by calling SendStream::finish on outstanding streams and waiting for the corresponding StreamEvent::Finished event.

If Streams::send_streams returns 0, all outstanding stream data has been delivered. There may still be data from the peer that has not been received.

pub fn datagrams(&mut self) -> Datagrams<'_>

Control datagrams

pub fn stats(&mut self) -> ConnectionStats

Returns connection statistics

pub fn path_stats(&mut self, path_id: PathId) -> Option<PathStats>

Returns path statistics

pub fn ping(&mut self)

Ping the remote endpoint

Causes an ACK-eliciting packet to be transmitted on the connection.

pub fn ping_path(&mut self, path: PathId) -> Result<(), ClosedPath>

Ping the remote endpoint over a specific path

Causes an ACK-eliciting packet to be transmitted on the path.

pub fn force_key_update(&mut self)

Update traffic keys spontaneously

This can be useful for testing key updates, as they otherwise only happen infrequently.

pub fn crypto_session(&self) -> &dyn Session

Get a session reference

pub fn is_handshaking(&self) -> bool

Whether the connection is in the process of being established

If this returns false, the connection may be either established or closed, signaled by the emission of a Connected or ConnectionLost event respectively. Note that locally-initiated closes via close() do not emit a ConnectionLost event.

For an established connection this essentially means the handshake is completed, but not necessarily yet confirmed.

pub fn is_closed(&self) -> bool

Whether the connection is closed

Closed connections cannot transport any further data. A connection becomes closed when either peer application intentionally closes it, or when either transport layer detects an error such as a time-out or certificate validation failure.

A ConnectionLost event is emitted with details when the connection is closed by the peer or due to an error. When the local application closes the connection via close(), no ConnectionLost event is emitted; instead, pending operations fail with ConnectionError::LocallyClosed.

pub fn is_drained(&self) -> bool

Whether there is no longer any need to keep the connection around

Closed connections become drained after a brief timeout to absorb any remaining in-flight packets from the peer. All drained connections have been closed.

pub fn accepted_0rtt(&self) -> bool

For clients, if the peer accepted the 0-RTT data packets

The value is meaningless until after the handshake completes.

pub fn has_0rtt(&self) -> bool

Whether 0-RTT is/was possible during the handshake

pub fn has_pending_retransmits(&self) -> bool

Whether there are any pending retransmits

pub fn side(&self) -> Side

Look up whether we’re the client or server of this Connection

pub fn path_observed_address( &self, path_id: PathId, ) -> Result<Option<SocketAddr>, ClosedPath>

Get the address observed by the remote over the given path

pub fn rtt(&self, path_id: PathId) -> Option<Duration>

Current best estimate of this connection’s latency (round-trip-time)

pub fn congestion_state(&self, path_id: PathId) -> Option<&dyn Controller>

Current state of this connection’s congestion controller, for debugging purposes

pub fn set_max_concurrent_streams(&mut self, dir: Dir, count: VarInt)

Modify the number of remotely initiated streams that may be concurrently open

No streams may be opened by the peer unless fewer than count are already open. Large counts increase both minimum and worst-case memory consumption.

pub fn set_max_concurrent_paths( &mut self, now: Instant, count: NonZeroU32, ) -> Result<(), MultipathNotNegotiated>

Modify the number of open paths allowed when multipath is enabled

When reducing the number of concurrent paths this will only affect delaying sending new MAX_PATH_ID frames until fewer than this number of paths are possible. To actively reduce paths they must be closed using Connection::close_path, which can also be used to close not-yet-opened paths.

If multipath is not negotiated (see the TransportConfig) this can not enable multipath and will fail.

pub fn max_concurrent_streams(&self, dir: Dir) -> u64

Current number of remotely initiated streams that may be concurrently open

If the target for this limit is reduced using set_max_concurrent_streams, it will not change immediately, even if fewer streams are open. Instead, it will decrement by one for each time a remotely initiated stream of matching directionality is closed.

pub fn set_send_window(&mut self, send_window: u64)

See TransportConfig::send_window()

pub fn set_receive_window(&mut self, receive_window: VarInt)

See TransportConfig::receive_window()

pub fn is_multipath_negotiated(&self) -> bool

Whether the Multipath for QUIC extension is enabled.

Multipath is only enabled after the handshake is completed and if it was enabled by both peers.

pub fn handle_network_change( &mut self, hint: Option<&dyn NetworkChangeHint>, now: Instant, )

Handle a change in the local address, i.e. an active migration

In the general (non-multipath) case, paths will perform a RFC9000 migration and be pinged for a liveness check. This is the behaviour of a path assumed to be recoverable, even if this is not the case.

Clients in a connection in which multipath has been negotiated should migrate paths to new PathIds. For paths that are known to be non-recoverable can be migrated to a new PathId by closing the current path, and opening a new one to the same remote. Treating paths as non recoverable when necessary accelerates connectivity re-establishment, or might allow it altogether.

The optional hint allows callers to indicate when paths are non-recoverable and should be migrated to new a PathId.

pub fn current_mtu(&self) -> u16

Storage size required for the largest packet that can be transmitted on all currently available paths

Buffers passed to Connection::poll_transmit should be at least this large.

When multipath is enabled, this value is the minimum MTU across all available paths.

pub fn add_nat_traversal_address( &mut self, address: SocketAddr, ) -> Result<(), Error>

Add addresses the local endpoint considers are reachable for nat traversal.

pub fn remove_nat_traversal_address( &mut self, address: SocketAddr, ) -> Result<(), Error>

Removes an address the endpoing no longer considers reachable for nat traversal

Addresses not present in the set will be silently ignored.

pub fn get_local_nat_traversal_addresses( &self, ) -> Result<Vec<SocketAddr>, Error>

Get the current local nat traversal addresses

pub fn get_remote_nat_traversal_addresses( &self, ) -> Result<Vec<SocketAddr>, Error>

Get the currently advertised nat traversal addresses by the server

pub fn initiate_nat_traversal_round( &mut self, now: Instant, ) -> Result<Vec<SocketAddr>, Error>

Initiates a new nat traversal round

A nat traversal round involves advertising the client’s local addresses in REACH_OUT frames, and initiating probing of the known remote addresses. When a new round is initiated, the previous one is cancelled.

For all probes that succeed, if any, a new path will be opened on the successful 4-tuple.

Returns the server addresses that are now being probed. If addresses fail due to spurious errors, these might succeed later and not be returned in this set.